Merged PR 30416: Don't use type comparison to select customer attribute dropdown value
## What's being changed

In this PR we are removing the type comparison when fetching the selected value of dropdown-format custom customer attribute, during customer sync.

## Why it's being changed

In some cases (e.g. when you create the attributes programmatically) data is stored as a type other than string, it then breaks the type comparison and fails to sync the value to EC for the selected attribute. Note that `$this->model->getData($attributeCode)` always returns a string.

## How to review / test this change

- Install this module https://github.com/amitshree/magento2-account-approval
- A new dropdown customer custom attribute will be created with label "Approve Account"
- Map this attribute and run contact sync
- Confirm that you get the expected value in EC

Related work items: #147907