Merged PR 30922: Add a new cron cleaner CLI task
## What's being changed

In this PR we added the ddg cleaner task to our cli task list

## Why it's being changed

This can be useful for merchants to run manually. The cleaner cron only runs once per month, so if that is missed for whatever (unlikely) reason, files and database rows can mount up.

## How to review / test this change

- Run bin/magento dotdigital:task and confirm that "Cleaner" will be returned to the list
- Select "Cleaner" [0] and confirm that works
- Make sure that cleaner task still works via cron

Related work items: #149595