Merged PR 31447: Check existing abandoned cart automations by website
## What's being changed

In this PR we are adding a website condition to the bit of code that asks whether an abandoned cart automation already exists (inside of a time window) for a given email address.

## Why it's being changed

If a customer drops a cart on multiple websites in the same Magento, we don't want to accidentally not enrol the contact on website 2 if they've already been enrolled for website 1.

## How to review / test this change

- Enable cart limit in Stores > Dotdigital> Configurations
- Enable AC program for two (or more) websites
- Login to website A and drop a cart
- Login to website B (with same email address) and drop a cart
- Make sure that we have automation programs for the same email address for both websites

Related work items: #151647