Merged PR 31463: Prevent duplicate contact rows when migrating subscribers
## What's being changed

We changed our logic around inserting and updating subscribers in email_contact during the install process.

## Why it's being changed

If there was an email address associated with a customer, and also with a subscription, BUT newsletter_subscriber did not have the customer_id relation (i.e. because the user subscribed whilst not logged in) - we ended up with duplicate rows in email_contact: one row for the customer, and another for the subscriber. This had a knock-on impact in not correctly syncing subscriber status for customers.

The `migrate` script now mirrors the behaviour of the connector when a user subscribes, i.e. if we find a matching record for the email address and the current scope, we update it.

## How to review / test this change

- Create a customer but do not subscribe
- Sign out
- With the same email address, sign up for the newsletter
- Note that you will have 1 row now in email_contact (the original row was updated)
- Run dotdigital:migrate
- Make sure that now we have a single record for that contact in email_contact table
- Sign up for the newsletter on website 2 with the same email address.
- Re-run dotdigital:migrate
- Make sure we have two rows in email_contact - a customer with is_subscriber = 1, and a subscriber with customer_id = 0.

Related work items: #151751