Merged PR 31451: Unsubscribe by account
The `unsubscribe()` bit of Subscriber sync. The code has been refactored to:

- move unsubscribe logic into its own class
- remove code from the Contact resource model

The unsubscribe flow now loops through connected accounts, on the basis that if a contact is suppressed for a particular account, Magento should unsubscribe that contact from all the websites which are connected to that account. There was some logic that created this enabled account data set in the Importer, so I moved that to its own class (it should be used in Order sync when we get to refactoring that).

The `getSuppressedContacts()` method has been removed from the Data helper, continuing a long-standing goal to reduce this massively bloated class.

I've removed a unit test that tested the Contact resource model, but have created a new one for the Unsubscriber class that has the same (if not better) coverage.

The main goal here was to fix a bug, where contacts using the same email address on different websites (connected to different EC accounts) would get unsubscribed from both websites, even if they were only suppressed on one EC account.

- Connect website 1 to EC account 1
- Connect website 2 to EC account 2
- On website 1, subscribe a guest.
- On website 2, subscribe a guest.
- Run subscriber sync.
- Suppress the contact in EC account 1.
- Run subscriber sync.
- Only the website 1 contact should be marked as unsubscribed in our `email_contact` table.
- Only the website 1 contact should be marked as unsubscribed in the core `newsletter_subscriber` table.

_Additional tests_

- Connect website 1 to EC account 1
- Connect website 2 to EC account 2
- Connect website 3 to EC account 2
- On website 2, subscribe a guest.
- On website 3, subscribe the same guest.
- Run subscriber sync.
- Suppress the contact in EC account 2.
- Run subscriber sync.
- The contact should be unsubscribed in website 2 and 3.
- The contact should be unsubscribed for all related store ids belonging to websites 2 and 3.

Related work items: #151649