Merged PR 32258: Pending Contact Updater: check for a $contact and expire if false
## What's being changed

Our PendingContactUpdater class logic.

## Why it's being changed

It's possible that a contact can sign up, trigger an automation, and be synced to EC as PendingOptIn - but be suppressed at the next check. If so, we need to expire the pending automation.

This scenario was causing automation sync and abandoned cart sync to break because of the error:
```
Trying to get property 'status' of non-object
```

## How to review / test this change

- Set your EC to enforce double opt in for all signups
- Set up subscriber automation
- Sign up an email address to trigger a new automation row
- Run automation sync
- Row should be marked as 'PendingOptIn'
- Now, mark the contact as suppressed in EC
- Re-run automation sync after 30 mins
- The automation should be marked as 'Expired'
- Retest for abandoned cart sync

Related work items: #154867