Merged PR 32261: Handle unsubscribed or suppressed contacts when processing PendingOptIn
## What's being changed

The logic that checks PendingOptIn contacts in abandoned cart sync and automation sync.

## Why it's being changed

It's possible that a contact can drop a cart, and be synced to EC as PendingOptIn - but be suppressed or unsubscribed at the next check. If so, we need to expire the pending abandoned cart.

This scenario was causing abandoned cart sync to break because of the error:
```
Trying to get property 'status' of non-object
```

Although the specific issue raised concerns abandoned cart sync, it will also affect automation sync, so this has also been fixed.

## How to review / test this change

- Set your EC to enforce double opt in for all signups
- Set up abandoned cart email series
- Drop a cart to trigger a new AC row
- Run abandoned cart sync
- Row should be marked as 'PendingOptIn'
- Now, mark the contact as suppressed in EC
- Re-run abandoned cart sync after 30 mins
- The row should be marked as 'Expired'
- Retest for automation sync

## Notes

This is a hotfix on master. In develop currently, this code has changed. We have improved the handling of pending contacts i.e. both syncs share the same logic, and automation sync uses the same date check to expire rows as abandoned cart sync (whereas right now, on master, pending contacts who never confirm will be re-checked indefinitely). This update will be available in 4.14.0.

Related work items: #154867