Merged PR 32466: Fix handling of parent product ids in bulk catalog update
## What's being changed

Three classes - ParentFinder, UpdateCatalog and UpdateCatalogBulk.

## Why it's being changed

The method `ParentFinder:: getConfigurableParentsFromProductIds()` was returning an array of product data, and we were trying to merge that array with an array of ids, and pass the combined array into an update query.

This is a regression introduced [here](https://github.com/dotmailer/dotmailer-magento2-extension/commit/ed1fb1629ec761f96f0d9284e904250f31caee5f) in 4.12.0.

While `UpdateCatalog` had a method to sort the ids out from the data arrays, `UpdateCatalogBulk` didn't, and we missed this.

## How to review / test this change

_Bulk import_
- Re-run the tests listed on this change for catalog imports: https://github.com/dotmailer/dotmailer-magento2-extension/commit/ed1fb1629ec761f96f0d9284e904250f31caee5f

_Product save after_
- Save a configurable child product in the admin
- Ensure that the child and the parent are marked as unprocessed

_Order sync_
- Place an order for a configurable product and others
- Run order sync
- Ensure that all products in the order items (parent and child for the configurable) are marked as unprocessed

Related work items: #155677