Merged PR 32457: Handle missing properties when fetching campaign
## What's being changed

We're adding some extra code to handle missing properties in the 'prepared-for-transactional-email' campaign object returned from dotdigital.

## Why it's being changed

It's not clear why expected properties might be missing. We've seen this error, during template sync, on different instances:
```
2021-09-22 13:00:54] report.CRITICAL: Notice: Undefined property: stdClass::$name in /app/vendor/dotmailer/dotmailer-magento2-extension/Model/Email/Template.php on line 394
```

This PR checks for the properties we need in order to update the saved template, and if any is missing, logs an error with the full object dumped (as an array) to try and understand what's going on.

## How to review / test this change

- To simulate the error, add something like this to the top of the `saveTemplate()` method:
```
$array = [
    'chaz' => 'foo'
];
$dmCampaign = (object) $array;
```
- Check the log for the error
- Ensure the template is returned and not updated with bogus or empty data

Related work items: #155633