Merged PR 32518: Delete API endpoint for scope if use default is checked
## What's being changed

In this PR when we saving new api creds we check if "use default" is checked for API username or API password.
If "use default" is checked then we delete the API endpoint for the current scope.

## Why it's being changed

There are a some cases that merchants on their first installation use a different account region than they use later. So if later on they decide to use for any of their website the default API account that belongs to a different region that means that they will end up having wrong api endpoint for the current scope.

This can make (some if not all) api calls will not work. Also this issue is never obvious for support because there is no relevant error message.

## How to review / test this change

- In a multiwebsite instance initialize your website 1 with an r2 account
- In default level save r1 account credentials
- Go to website 1 accounts section and click on "use default" for the API user creds.
- Go go to address books section (for website 1) and make sure that address books are loaded (that means api works).
- Make sure that api endpoint removed for website 1 as now uses the default api creds.
- Repeat same steps on develop, you will not be able to see any address books in dropdown for website 1

**Notes**
I've seen this issue happening a few times.
here is an example
https://dotmailer.zendesk.com/agent/tickets/660775

Related work items: #153871