Merged PR 32371: Restore guest carts outside of original session
## What's being changed

This PR makes it possible for a guest to receive an abandoned cart email on any device, and to reach a hydrated basket in Magento.

If the guest follows the getbasket link, and there is no current session or session quote, we simply set the quote id, and Magento does the the rest.

If the guest e.g. starts browsing on desktop, triggers the AC, then continues browsing on mobile, adds further (or different) items to a new quote, and then clicks the campaign getbasket link, we check for any items from the dropped basket that aren't present in the new basket, and add them.

## Why it's being changed

It's always been a frustration for merchants to be sending out AC email to (identified) guests, only for the recipients to land at an empty cart in Magento. This change should improve conversion rate for these shoppers.

There has been some confusion around this in the past. We've had a method called `checkMissingAndAdd()`, but it was only used for customers, and only if the customer's current quote id was different from the original abandoned one. This is probably more of an edge case than the guest scenario. In any case, it probably didn't work - because `$currentQuote->addItem($item)` only works if that item doesn't already have an id, which it would have, because it had already been attached to a different quote (this has been fixed as well).

## How to review / test this change

_Setup_
- Configure abandoned cart series for guests

_New session_
- Drop a cart as a guest
- Step through to the checkout, and enter your email address (this will trigger an onBlur event that captures your email address)
- Receive an AC email
- Use an incognito window to follow the getbasket link
- See your original cart items appear in a new basket

_Existing session with quote_
- Use an incognito window (or a different browser ideally) to create a new quote with some items
- In that window, paste the getbasket link from the campaign you received
- Your cart should be updated to contain the original items plus the items from the second quote

## Notes

For customers who drop a cart, behaviour is unchanged.
- Customers who drop a cart, and who then browse as a guest on a different device (and add other, potentially different, products to the cart) will see the guest cart _combined_ with the stored quote items when they next log in. This is standard Magento behaviour.
- We do run `checkMissingAndAdd()` for customers but _only_ if:
    - they are logged in already at the point they hit getbasket
    - they have a quote but its ID _isn't_ the same as the ID in the URL, even though both share the same customer_id (I don't see how this is possible, perhaps it was in some earlier version)
- The stored quote that we request via the getbasket EDC will always be the latest 'live' basket. Even if the abandoned cart email shows two products, but one of those was removed from the cart after the campaign was sent, getbasket will only retrieve the produc...