Merged PR 33183: Update region prefix pattern to accept API endpoint changes
## What's being changed

In this PR we updated the regex we use to return the region prefix so to accept any domain

## Why it's being changed

The Dotdigital API endpoint is (at some point) changing from api.dotmailer.com to api.dotdigital.com.

If regex remains as it is "dotdigital" domain will be rejected as a result will fail to fetch the region prefix.

## How to review / test this change

- use this tool https://www.functions-online.com/preg_match.html to validate that the updated regex will accept endpoints like "https://r2-api.dotdigital.com"

Related work items: #157440