Merged PR 33686: Apply data patches and remove old install-upgrade scripts
## What's being changed

In this PR we removed the old type install/upgrade data scripts and replaced them with data patches.

For the DataMigration patch, this will do everything that the old InstallData script did, i.e. everything that `bin/magento dotdigital:migrate` does. However it is designed for fresh installs, and will only populate data in a table if that table is empty (so that merchants with populated tables won't have any alterations).

## Why it's being changed

This completes work already started to align our code with Magento's declarative schema. Simply having the install/upgrade scripts present is now enough to fail marketplace submission.

## How to review / test this change

- In a fresh connector installation confirm that all the data are being migrated with the correct format
- In existing installations - Confirm that data patches applied for old type of data (trying to simulate that)
- Confirm that we do not regenerate a new passcode

**Additional review steps after update**
- Test migrate command for empty and non empty tables
- Test migrate command for specific tables (bin/magento dotdigital:migrate --table=email_catalog)
- Confirm that the migrate script applies data installation **only for not empty tables**

## Notes

Currently our migration data script contains DDL transactions (delete and alter increments of each table) which throws an error when executed in patch script.

So in this PR we move (empty tables) out of the run() function which installs all the data.
- In migrate command we have to empty the selected table(s) before we run the migration.
- In the DataMigration patch we check each table if is empty or no before we proceed with the data installation.

Related work items: #159609