Merged PR 33540: Fix cURL option and value
## What's being changed

We've removed the setting of curl options for basic auth.

## Why it's being changed

`CURLAUTH_BASIC` is not a valid [cURL option](https://www.php.net/manual/en/function.curl-setopt.php), it's a value to be set on the `CURLOPT_HTTPAUTH` option. Seemingly this has either been changed in PHP 8+, or was not strictly enforced before.

So the correct way of writing this would have been:
```
curl_setopt($ch, CURLOPT_HTTPAUTH, CURLAUTH_BASIC);
```

However, this is redundant. cURL uses basic auth by default when a `CURLOPT_USERPWD` is passed.

## How to review / test this change

- Test API validation in PHP 7.x and 8.x
- Ensure no errors are thrown and that valid creds succeed

Related work items: #159382