Merged PR 33946: Handle passcode generation after table migration
## What's being changed

We're moving the code that generates a new dynamic content passcode to run after the main tables migration, as opposed to in the `run()` function itself.

## Why it's being changed

When creating the `MigrateData` patch it became apparent that `generateAndSaveCode()` would be run for each table in the loop.

In fact this method only needs to be run in the context of a full `dotdigital:migrate` migration (without a table supplied as an argument). As before the method won't overwrite an existing stored passcode.

## How to review / test this change

- Delete the stored passcode
- Re-run `dotdigital:migrate` with a table supplied
- The passcode will not be created
- Re-run `dotdigital:migrate` with no table supplied
- The passcode will be created
- Re-run to confirm it is still not overwritten
- Remove the patch from the table and upgrade to re-apply the patch

Related work items: #160394