Merged PR 34049: Convert email to lowercase before checking array index
## What's being changed

We're converting an email string to lowercase before using its value as an array index.

## Why it's being changed

Array indices are case-sensitive in PHP, and Dotdigital converts email addresses to lowercase before storing them. If a contact subscribes with newContact2@emailsim.io, newcontact2 is stored in Dotdigital. In the resubscribe flow, this caused an undefined index error when trying to find a match on `$contacts['newContact2@emailsim.io']`.

## How to review / test this change

- Subscribe newContact2@emailsim.io
- Sync to Dotdigital
- Unsubscribe in Magento (but don't re-sync or run the importer)
- Run the resubscribe cron
- newContact2@emailsim.io should be resubscribed

Related work items: #161443