Merged PR 34231: Add bulk actions to set products as processed / unprocessed for catalog sync
## What's being changed

We've added MassActions in the catalog report to bulk set products as processed or unprocessed. Setting all products to unprocessed is equivalent to a catalog reset.

## Why it's being changed

This accompanies the forthcoming improvements to catalog sync. If we encounter problems exporting one particular product in a batch, we'll skip it, log the exception, but continue batching. The product will be marked as processed, but will be referenced by id in the logs, so now we can manually mark this product as 'Not processed' so it will be re-synced.

There are other use cases of course - this removes the need to 'edit' a product in order to have it re-sync. It also allows admins to mark whole chunks of their catalog as processed and not sync these items, which can be useful if these products have some sort of known data issue.

## How to review / test this change

- Select one item from the Catalog report
- Change its status from Processed to Not Processed
- Repeat with a page of items, and with the entire collection

Related work items: #162006