Merged PR 31571: Document CLI commands in README
## What's being changed

We're adding some developer documentation for the module's CLI commands. We've also moved the release history to a dedicated CHANGELOG.md.

## Why it's being changed

The CLI commands started as utility commands for our own development purposes, but increasingly they are used by merchants: the `sync` command is handy when debugging cron issues, and `migrate` is regularly used either to reset data, when deploying, or at go-live.

Related work items: #151271