Merged PR 34629: Only use getTo if EmailMessageInterface is available
## What's being changed

This PR skips the reference to `getTo()` if that method is unavailable.

## Why it's being changed

Only the `EmailMessageInterface` (2.3.3+) has a `getTo()` method, so in older Magento versions this method call was throwing an exception. In fact the invalid email address is exposed anyway in the stack trace, so this was never really essential. Either way this will no longer cause an issue.

## How to review / test this change

- In Magento 2.3.2, enable the connector and transactional SMTP
- Try to subscribe an invalid email address like test@paok.paok
- Check the error log line and confirm no further errors occur

Related work items: #163101