Merged PR 34537: Update catalog reset if catalog image type changed
## What's being changed

We've removed an `ImageTypes` observer that was introduced in 4.10.0.

## Why it's being changed

This observer was supposed to reset the catalog if the 'Catalog sync' image type value changed (but not if any other value changed in **Dotdigital > Configuration**. This surely worked when we implemented it, but then I started to notice random catalog resets.

It seems like all the page paths come back in the 'changed paths' array (certainly since 2.3.5, which is as far back as I've tested), so we were resetting catalog any time any value was changed on this page.

Moving the logic to the backend model fixes this.

## How to review / test this change

- Go to Dotdigital > Configuration
- Save the page with no changes (catalog is not reset)
- Save the page with changes to any other fields apart from Catalog sync image type (catalog is not reset)
- Save the page with changes to Catalog sync image type (catalog is reset)

Related work items: #162933