Merged PR 34911: Pass $selectStatement to the update where clause
## What's being changed

I've changed the method signature of a method in `UpdateTypeInterface`, one of the three types of data migration we perform when populating Dotdigital module tables.

## Why it's being changed

The AbstractDataMigration says:
```
if ($this instanceof UpdateTypeInterface) {
    $this->update($this->getSelectStatement());
}
```

I think we ought to pass this `Select` down, in the `update()` method, to `$this->getUpdateWhereClause($selectStatement)`, in exactly the same way as we do in `batchInsert()`.

## How to review / test this change

- Re-run bin/magento dotdigital:migrate
- Pay attention to the function of the `UpdateContactsWithSubscriberCustomers` script (you'll need a Magento customer who is also a subscriber to verify)

Related work items: #164266