Merged PR 35253: Skip catalog batch if it contains no products
## What's being changed

We're checking a batch contains products before attempting to queue it with the importer.

## Why it's being changed

If a mega batch contains exported catalogs for more than one website, but one of the websites exported no products, we will get a failure logged e.g.:
```
dotdigital.INFO: Failed to register with IMPORTER.Type(Catalog) / Scope(Bulk) / Website(2) / Product Ids() [] []
```
This is misleading and not necessary.

## How to review / test this change

- Start with a fully synced catalog
- Set one product to Not Processed which you know to not exist in all scopes
- Run catalog sync
- Do not see a 'Failed to register' log line.

Related work items: #165476