Merged PR 35019: Do not use offset for the customer sales migration type
## What's being changed

The `InsertEmailContactTableCustomerSales` migration.

## Why it's being changed

We're setting the offset flag to false for this query. Without this, the offset means we miss 500 records on each iteration, and consequently only migrate 50% of the data.

Because the query is NOT IN email_contact, on the second iteration of the loop, we start with an offset of +500. Then select statement finds 500 fewer anyway, because 500 records have already been migrated already, on the first iteration. So it actually does a double offset - 500 rows affected plus 500 offset, i.e. 2x the batch size, so 50% of the data overall.

## How to review / test this change

- Set up performance fixtures so you have over 500 customer orders (you may need Fanis's new customer orders fixture for this)
- Turn on Global Account Sharing (this migration only runs in GAS mode)
- Run bin/magento dotdigital:migrate
- Confirm that all expected customers are populated in the email_contact table

Related work items: #164802