Merged PR 35759: Refactor inheritance in data migration
## What's being changed

The AbstractDataMigration throws errors in PHPStan because it calls methods only defined in its child classes e.g. `getUpdateBindings` (which is used in Update and BulkUpdate migration types, but not in Insert types).

Currently all the migration types implement an Interface, which is fine, but if a new sub-class were to be created that did not implement the interface, we would get a fatal error if the class called `update()` but did not have an implementation of the `getUpdateBindings` method. Admittedly, this is unlikely, but PHPStan is flagging a bit of a code smell here.

See discussion on this ticket: https://github.com/phpstan/phpstan/issues/1323

## Why it's being changed

I've restructured things so that we have an AbstractDataMigration, with methods and abstract methods common to all migration types, PLUS abstract classes for the three data migration types: update, bulk update and insert.

## How to review / test this change

- Re-run `bin/magento dotdigital:migrate` for a large dataset
- Compare the numbers of rows migrated before and after these changes
- Confirm PHPStan and coding standards builds are green for the affected files

## Notes

TODO: PR for B2B

Related work items: #167675