Merged PR 35965: Update minimum PHP requirement to 7.2
## What's being changed

The minimum PHP requirement for using our Magento connectors is now 7.2.

## Why it's being changed

PHP 7.2 is supported for all 2.3+ versions, so this doesn't affect the Magento versions we support. However, even PHP 7.3 is now end of life. This is a practical step to reduce our exposure to errors now that we also support PHP 8.1.

## How to review / test this change

- Switch to PHP 7.2
- Run `composer update`
- Run `bin/magento setup:upgrade`

Related work items: #167169