Merged PR 35866: Do not send cart insight data if any exception is thrown when gathering payload
## What's being changed

Cart insight data payload.

## Why it's being changed

It has emerged from at least one support ticket that `quote_item.name` can be empty in a cart insight data line item. If this happens, it can break segmentation for cart insight data in Dotdigital.

Historically we have added exception handling in this area in a way that skipped anything bad but tried to send as much as possible. However we've now decided that incomplete data is worse than no data.

Whereas previously we would skip all line items (but send the totals) if there was an error getting line items, and skip individual items if there was an error fetching data for that particular line item, this PR goes further. No cart insight data will now be sent at all if an exception of any kind was thrown when getting the data payload. This prevents missing data from breaking things on Dotdigital, and also prevents missing line items from skewing the data (e.g. if only 3 out of 4 line items are sent in the payload, the grand total will not match the sum of the item values).

## How to review / test this change

- Drop a cart on the storefront containing multiple items
- Go into the `catalog_product_entity_varchar` table and manually delete the name value of one of the products in your cart (you'll need to search for the matching entity_id or row_id, and if it's configurable product, you need to delete the name of the parent product)
- Run AbandonedCart sync from the CLI
- See an exception logged with a relevant message
- No cart insight data should be sent

_Additional tests_

- Manually throw an exception (`throw new \Exception('Exception thrown in <method>');` in `Quote::getAllVisibleItems()`
- Manually throw an exception in `UrlFinder::fetchFor()`
- In both cases, the exception should be caught at the same place and no cartinsight data should be sent.

## Notes

When debugging AbandonedCart sync, and cart insight, use program enrolment (it's easier to run) and manually modify the 'from' and 'to' dates in Model/AbandonedCart/ProgramEnrolment/Interval.php to create a custom interval that catches your quote's `updated_at` value.

`$item->getName()` in fact does not pull the value from `quote_item.name`, but instead from the matching attribute value in `catalog_product_entity_varchar`. This can't be deleted from the front end (name is a required field) but only via the db.

Related work items: #168116