Merged PR 36158: Only run StockUpdatePlugin for Web API requests
## What's being changed

The StockUpdatePlugin.

## Why it's being changed

The StockUpdatePlugin was introduced to meet a very specific need - one merchant was using a Web API request to modify stock levels, and needed a way to re-sync products from our table off the back of this. We didn't make it specific enough, so created a service class to ensure that we didn't update our table twice (once for the plugin and once for the observer).

Recently we considered removing our `catalog_product_save_after` observer, on the grounds that we were duplicating code. However the observer still feels like a reliable and intuitive approach to updating products when they are saved. Plus, StockRegistryInterface was deprecated by 2.3.4 (if not before) so who knows what will happen to it in future versions.

This PR uses Magento\Framework\App\State to check in the StockUpdatePlugin whether the incoming request is a web API request. If there are stock updates that somehow don't originate in a web API request, they will still not trigger a double update thanks to the catalog service class.

## How to review / test this change

- Create an integration in System > Integrations, activate it and make a note of the access token
- Set up Postman (Authorization: Bearer Token, token = your access token) to make an API call to /rest/default/V1/products/24-MB01/stockItems/1 [pick any SKU, but find the item id for it in the `cataloginventory_stock_item` table].
- Check the relevant product id is processed = 1 currently, in `email_catalog`
- Put the following data in the body:
```
{
  "stockItem": {
    "item_id": 1,
    "product_id": 1,
    "stock_id": 1,
    "qty": 96,
    "is_in_stock": true,
    "is_qty_decimal": false,
    "use_config_min_qty": true,
    "min_qty": 0,
    "use_config_min_sale_qty": 0,
    "min_sale_qty": 0,
    "use_config_max_sale_qty": true,
    "max_sale_qty": 0,
    "use_config_backorders": true,
    "backorders": 0,
    "use_config_notify_stock_qty": true,
    "notify_stock_qty": 0,
    "use_config_qty_increments": true,
    "qty_increments": 1.0000,
    "use_config_enable_qty_inc": true,
    "enable_qty_increments": true,
    "use_config_manage_stock": true,
    "manage_stock": true,
    "low_stock_date": "string",
    "is_decimal_divided": false,
    "stock_status_changed_auto": 0,
    "extension_attributes": {}
  }
}
```
- Assuming you get a successful response, confirm the item is processed = 0 in our table
- Re-test saving products in the admin and confirm items are again set to processed = 0.

## Notes

I did look into whether we could use the hasDataChanges() method on the product model to see if it actually had been changed - however currently this always returns true, and is an active issue: https://github.com/magento/magento2/issues/34920.

Related work items: #169511