Merged PR 36320: Remove multi source inventory from core module
## What's being changed

Reverted stock finder model to not depend on Magento Multi Source Inventory (introduced in 4.14.0).

## Why it's being changed

Merchants would prefer to disable the Magento MSI modules when they do not require them for performance and overhead reasons, Dotdigital StockFinder relies on the modules as part of the core. To remedy this we created a new extension (Dotdigitalgroup_Inventory) that can be used to handle the required MSI functions.

This allows merchants to disable the Dotdigital Inventory module if they are not using MSI.

## How to review / test this change

1. See [this ticket](https://dev.azure.com/dotdigital/ec/_sprints/taskboard/Error%20Unknown/ec/2022/Sprint%206?workitem=169911) for removal steps
2. Run php7.4 bin/magento setup:upgrade
3. Edit the email catalog table marks a couple of rows processed column to 0
4. Run catalog sync
5. Check the processed columns in the email catalog table are now 1

## Notes

Merchants on 4.14.0-4.15.0 who have been running into errors because this should:
- Upgrade their metapackage
- Run `bin/magento module:disable Dotdigitalgroup_Inventory`

Related work items: #169902, #169903