Merged PR 36298: Do not save Chat API space id and token at trial account creation
## What's being changed

In this PR we're removing any code that stores the Chat API space id and token in Magento after a trial account is created via the microsite.

We also fixed a small bug affecting instances with a port specified in the URL.

## Why it's being changed

Free chat account creation is being withdrawn, and we're standardising the trial onboarding routes.

## How to review / test this change

- Switch to the new code for your local microsite
- Register for a chat trial from Magento
- Confirm account creation works as expected and that API creds are passed to and stored in Magento

Related work items: #170164